changed routes and removed unnecessary files
